beta tester improvements(file consolidation, add email invite
