+++ server/utils/requireSuperadmin.ts
@@ -2,8 +2,8 @@ import { serverSupabaseUser, serverSupabaseServiceRole } from '#supabase/server'
 import type { H3Event } from 'h3'
 
 /**
- * Ensures the current request is from an authenticated admin user.
- * Throws 401 if not authenticated, 403 if not admin.
+ * Ensures the current request is from an authenticated platform superadmin (is_superadmin).
+ * Throws 401 if not authenticated, 403 if not superadmin.
  * Returns the authenticated user row from the `users` table.
  */
 export async function requireSuperadmin(event: H3Event) {
@@ -17,7 +17,7 @@ export async function requireSuperadmin(event: H3Event) {
 
   const { data: dbUser, error } = await client
     .from('users')
-    .select('id, email, is_admin')
+    .select('id, email, is_superadmin')
     .eq('id', authUser.id)
     .single()
 
@@ -25,7 +25,7 @@ export async function requireSuperadmin(event: H3Event) {
     throw createError({ statusCode: 401, statusMessage: 'User not found' })
   }
 
-  if (!dbUser.is_admin) {
+  if (dbUser.is_superadmin !== true) {
     throw createError({ statusCode: 403, statusMessage: 'Superadmin access required' })
   }
