+++ types/index.ts
@@ -302,6 +302,9 @@ export interface BetaTester {
   added_by: string | null
   created_at: string
   updated_at: string
+  first_signed_in_at: string | null
+  last_signed_in_at: string | null
+  registered_at: string | null
 }
 
 // ============================================================================
